Update plugin to support usage of MailMojo API
We now support retrieval of email list from the MailMojo API. Meaning
that one has to add an access token on the settings section. We still
support username and the old list ID, for now, so activated forms will
still work when one upgrades. The access token has to be retrieved from
the integration section in MailMojo.

Translations are missing and will be updated in a later patch when the
new features are in.