Add link to settings page from plugin actions
This also adds a link to the MailMojo plugin settings page from the
plugin actions. Plugin actions are the list of links displayed for each
plugin in the list of installed plugins. The settings link will be
prepended to the list of default actions, which seems a common practice
of other plugins like the built-in Akismet plugin.