Disable plugin for PHP versions older than 5.4.0
Our new plugin which uses the MailMojo API SDK has a minimum requirement
of PHP version 5.4.0 to work. For older PHP versions the plugin will
simply crash on syntax errors.

This adds a version check on initial boot of the plugin which will
prevent any of the incompatible code to load if the running PHP version
is older than 5.4.0. In addition we make it impossible to activate the
plugin in WordPress for these older PHP versions, and display an error
message if that is attempted.

The PHP version requirement is also mentioned in the changelog so users
can be aware before attempting to install.