Fix syntax error in PHP 5.4
There was still one syntax usage that is only supported from PHP 5.5+,
giving `Fatal error: Can't use method return value in write context` in
PHP 5.4. This is a simple fix of not passing a method return value
directly into the `empty()` construct, but rather via a variable.
Similarly we can't pass statically bound variables directly into
`empty()`.

This fixes it and bumps version to 1.0.1 for a patch release.