Validate access token on plugin settings page
When viewing the MailMojo plugin settings page we also perform an API
call to check if the token is valid when a token has been set. If the
API call fails we display an error informing of this and instructing to
fetch a new token from the MailMojo integrations page.

This means the token will be validated when first configuring the
plugin, giving the user more immediate feedback if the integration is
working.